Add QR code functionality to Ground Station
A QR code that users can scan to open a mapping app on their mobile
device is now displayed on the main screen of the Ground Station
program.

The last location received from the tracker is encoded in the
QR code and displayed on the map. The QR code is updated whenever a new
valid message is received.

This adds another dependency to the python program: pyqrcode.