Change serial port write_timeout to 5 seconds in Ground Station
This is a minor change to how the program writes to the serial port.

Before, the serial port would block indefinitely if a write was
unsuccessful (write_timeout=None). Now the serial port will display an
error and the program will continue if no data can be sent after 5
seconds (write_timeout=5).